+++ docs/source/changelog.md
@@ -5,6 +5,7 @@
 - Remove stage argument from [](#RawSql) initializer.
 - Add [](#RawSql) to public API.
 - Fix [](#PrefectTwoEngine) failing on retrieval of results.
+- Added [](#Flow.get_stage()), [](#Flow.get_task_from_string()) and [](#Stage.get_task()) methods.
 
 ## 0.6.0 (2023-07-07)
 - Added [`delete-schemas`](#reference/cli:delete-schemas) command to `pipedag-manage` to help with cleaning up database